requirements.txt
Flask==2.2.5
Flask-AppBuilder==4.3.7
Flask-APScheduler==1.13.1
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-Cors==4.0.2
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
PyJWT==2.9.0
pytest-flask==1.3.0
APScheduler==3.10.4
pytest-mock==3.14.0
Werkzeug==3.0.6
pydantic==2.9.2
pydantic_core==2.23.4
requests==2.32.2
pillow==9.1.1
Authlib==1.3.2
Flask-OAuthlib==0.9.6
gunicorn==23.0.0
psycopg2-binary==2.9.6
WTForms==3.1.2
pre_commit==4.0.1
black==24.10.0
flake8==7.1.1
isort==5.13.2
cryptography==44.0.2 
python-slugify==8.0.4
celery==5.4.0
redis==5.0.1


